Move the mouse support into its own files.
