+++ inputs/__init__.py
@@ -43,42 +43,37 @@ from __future__ import print_function
 from __future__ import division
 
 import os
-import io
 import glob
-import struct
-import time
 from warnings import warn
 import ctypes
+
 
 from .devices.led.led import LED, GamepadLED, SystemLED
 
 from .platforms.baselistener import BaseListener
-from .devices.gamepad._win import XinputState
-from .devices.device import InputDevice
 
 from .constants import (
     XINPUT_DLL_NAMES,
     XINPUT_ERROR_DEVICE_NOT_CONNECTED,
     XINPUT_ERROR_SUCCESS,
     WIN_KEYBOARD_CODES,
-    WIN_MOUSE_CODES,
-    MAC_EVENT_CODES,
     MAC_KEYS,
     EVENT_MAP,
     APPKIT_KB_PATH,
-    QUARTZ_MOUSE_PATH,
-    APPKIT_MOUSE_PATH,
 )
 
 from .errors import (
-    PERMISSIONS_ERROR_TEXT,
     UnknownEventType,
     UnknownEventCode
 )
 
 from .platforms.system import WIN, MAC, NIX
-from .platforms.c import DWORD, HANDLE, WPARAM, LPARAM, MSG, EVENT_FORMAT, EVENT_SIZE, convert_timeval
+from .platforms.c import DWORD, HANDLE, WPARAM, LPARAM, MSG
 from .devices.gamepad.gamepad import GamePad
+from .devices.device import InputDevice
+from .devices.gamepad._win import XinputState
+from .devices.mouse.mouse import Mouse, MightyMouse
+
 
 __version__ = "0.6"
 
@@ -110,25 +105,6 @@ class KBDLLHookStruct(ctypes.Structure):
     ]
 
 
-class MSLLHookStruct(ctypes.Structure):
-    """Contains information about a low-level mouse input event.
-
-    For full details see Microsoft's documentation:
-
-    https://msdn.microsoft.com/en-us/library/windows/desktop/
-    ms644970%28v=vs.85%29.aspx
-    """
-
-    # pylint: disable=too-few-public-methods
-    _fields_ = [
-        ("x_pos", ctypes.c_long),
-        ("y_pos", ctypes.c_long),
-        ("reserved", ctypes.c_short),
-        ("mousedata", ctypes.c_short),
-        ("flags", DWORD),
-        ("time", DWORD),
-        ("extrainfo", ctypes.c_ulong),
-    ]
 
 
 
@@ -202,544 +178,6 @@ def keyboard_process(pipe):
     """Single subprocess for reading keyboard events on Windows."""
     keyboard = WindowsKeyboardListener(pipe)
     keyboard.listen()
-
-
-class WindowsMouseListener(BaseListener):
-    """Loosely emulate Evdev mouse behaviour on Windows.  Listen (hook
-    in Windows terminology) for key events then buffer them in a pipe.
-    """
-
-    def __init__(self, pipe):
-        self.pipe = pipe
-        self.hooked = None
-        self.pointer = None
-        self.mouse_codes = WIN_MOUSE_CODES
-        super(WindowsMouseListener, self).__init__(pipe)
-
-    @staticmethod
-    def listen():
-        """Listen for mouse input."""
-        msg = MSG()
-        ctypes.windll.user32.GetMessageA(ctypes.byref(msg), 0, 0, 0)
-
-    def get_fptr(self):
-        """Get the function pointer."""
-        cmpfunc = ctypes.CFUNCTYPE(
-            ctypes.c_int, WPARAM, LPARAM, ctypes.POINTER(MSLLHookStruct)
-        )
-        return cmpfunc(self.handle_input)
-
-    def install_handle_input(self):
-        """Install the hook."""
-        self.pointer = self.get_fptr()
-
-        self.hooked = ctypes.windll.user32.SetWindowsHookExA(
-            14, self.pointer, ctypes.windll.kernel32.GetModuleHandleW(None), 0
-        )
-        if not self.hooked:
-            return False
-        return True
-
-    def uninstall_handle_input(self):
-        """Remove the hook."""
-        if self.hooked is None:
-            return
-        ctypes.windll.user32.UnhookWindowsHookEx(self.hooked)
-        self.hooked = None
-
-    def handle_input(self, ncode, wparam, lparam):
-        """Process the key input."""
-        x_pos = lparam.contents.x_pos
-        y_pos = lparam.contents.y_pos
-        data = lparam.contents.mousedata
-
-        # This is how we can distinguish mouse 1 from mouse 2
-        # extrainfo = lparam.contents.extrainfo
-        # The way windows seems to do it is there is primary mouse
-        # and all other mouses report as mouse 2
-
-        # Also useful later will be to support the flags field
-        # flags = lparam.contents.flags
-        # This shows if the event was from a real device or whether it
-        # was injected somehow via software
-
-        self.emulate_mouse(wparam, x_pos, y_pos, data)
-
-        # Give back control to Windows to wait for and process the
-        # next event
-        return ctypes.windll.user32.CallNextHookEx(self.hooked, ncode, wparam, lparam)
-
-    def emulate_mouse(self, key_code, x_val, y_val, data):
-        """Emulate the ev codes using the data Windows has given us.
-
-        Note that by default in Windows, to recognise a double click,
-        you just notice two clicks in a row within a reasonablely
-        short time period.
-
-        However, if the application developer sets the application
-        window's class style to CS_DBLCLKS, the operating system will
-        notice the four button events (down, up, down, up), intercept
-        them and then send a single key code instead.
-
-        There are no such special double click codes on other
-        platforms, so not obvious what to do with them. It might be
-        best to just convert them back to four events.
-
-        Currently we do nothing.
-
-        ((0x0203, 'WM_LBUTTONDBLCLK'),
-         (0x0206, 'WM_RBUTTONDBLCLK'),
-         (0x0209, 'WM_MBUTTONDBLCLK'),
-         (0x020D, 'WM_XBUTTONDBLCLK'))
-
-        """
-        # Once again ignore Windows' relative time (since system
-        # startup) and use the absolute time (since epoch i.e. 1st Jan
-        # 1970).
-        self.update_timeval()
-
-        events = []
-
-        if key_code == 0x0200:
-            # We have a mouse move alone.
-            # So just pass through to below
-            pass
-        elif key_code == 0x020A:
-            # We have a vertical mouse wheel turn
-            events.append(self.emulate_wheel(data, "y", self.timeval))
-        elif key_code == 0x020E:
-            # We have a horizontal mouse wheel turn
-            # https://msdn.microsoft.com/en-us/library/windows/desktop/
-            # ms645614%28v=vs.85%29.aspx
-            events.append(self.emulate_wheel(data, "x", self.timeval))
-        else:
-            # We have a button press.
-
-            # Distinguish the second extra button
-            if key_code == 0x020B and data == 2:
-                key_code = 0x020B2
-            elif key_code == 0x020C and data == 2:
-                key_code = 0x020C2
-
-            # Get the mouse codes
-            code, value, scan_code = self.mouse_codes[key_code]
-            # Add in the press events
-            scan_event, key_event = self.emulate_press(
-                code, scan_code, value, self.timeval
-            )
-            events.append(scan_event)
-            events.append(key_event)
-
-        # Add in the absolute position of the mouse cursor
-        x_event, y_event = self.emulate_abs(x_val, y_val, self.timeval)
-        events.append(x_event)
-        events.append(y_event)
-
-        # End with a sync marker
-        events.append(self.sync_marker(self.timeval))
-
-        # We are done
-        self.write_to_pipe(events)
-
-
-def mouse_process(pipe):
-    """Single subprocess for reading mouse events on Windows."""
-    mouse = WindowsMouseListener(pipe)
-    mouse.listen()
-
-
-class QuartzMouseBaseListener(BaseListener):
-    """Emulate evdev mouse behaviour on mac."""
-
-    def __init__(self, pipe):
-        super(QuartzMouseBaseListener, self).__init__(pipe, codes=dict(MAC_EVENT_CODES))
-        self.active = True
-        self.events = []
-
-    def _get_mouse_button_number(self, event):
-        """Get the mouse button number from an event."""
-        raise NotImplementedError
-
-    def _get_click_state(self, event):
-        """The click state from an event."""
-        raise NotImplementedError
-
-    def _get_scroll(self, event):
-        """The scroll values from an event."""
-        raise NotImplementedError
-
-    def _get_absolute(self, event):
-        """Get abolute cursor location."""
-        raise NotImplementedError
-
-    def _get_relative(self, event):
-        """Get the relative mouse movement."""
-        raise NotImplementedError
-
-    def handle_button(self, event, event_type):
-        """Convert the button information from quartz into evdev format."""
-        # 0 for left
-        # 1 for right
-        # 2 for middle/center
-        # 3 for side
-        mouse_button_number = self._get_mouse_button_number(event)
-
-        # Identify buttons 3,4,5
-        if event_type in (25, 26):
-            event_type = event_type + (mouse_button_number * 0.1)
-
-        # Add buttons to events
-        event_type_string, event_code, value, scan = self.codes[event_type]
-        if event_type_string == "Key":
-            scan_event, key_event = self.emulate_press(
-                event_code, scan, value, self.timeval
-            )
-            self.events.append(scan_event)
-            self.events.append(key_event)
-
-        # doubleclick/n-click of button
-        click_state = self._get_click_state(event)
-
-        repeat = self.emulate_repeat(click_state, self.timeval)
-        self.events.append(repeat)
-
-    def handle_scrollwheel(self, event):
-        """Handle the scrollwheel (it is a ball on the mighty mouse)."""
-        # relative Scrollwheel
-        scroll_x, scroll_y = self._get_scroll(event)
-
-        if scroll_x:
-            self.events.append(self.emulate_wheel(scroll_x, "x", self.timeval))
-
-        if scroll_y:
-            self.events.append(self.emulate_wheel(scroll_y, "y", self.timeval))
-
-    def handle_absolute(self, event):
-        """Absolute mouse position on the screen."""
-        (x_val, y_val) = self._get_absolute(event)
-        x_event, y_event = self.emulate_abs(int(x_val), int(y_val), self.timeval)
-        self.events.append(x_event)
-        self.events.append(y_event)
-
-    def handle_relative(self, event):
-        """Relative mouse movement."""
-        delta_x, delta_y = self._get_relative(event)
-        if delta_x:
-            self.events.append(self.emulate_rel(0x00, delta_x, self.timeval))
-        if delta_y:
-            self.events.append(self.emulate_rel(0x01, delta_y, self.timeval))
-
-    # pylint: disable=unused-argument
-    def handle_input(self, proxy, event_type, event, refcon):
-        """Handle an input event."""
-        self.update_timeval()
-        self.events = []
-
-        if event_type in (1, 2, 3, 4, 25, 26, 27):
-            self.handle_button(event, event_type)
-
-        if event_type == 22:
-            self.handle_scrollwheel(event)
-
-        # Add in the absolute position of the mouse cursor
-        self.handle_absolute(event)
-
-        # Add in the relative position of the mouse cursor
-        self.handle_relative(event)
-
-        # End with a sync marker
-        self.events.append(self.sync_marker(self.timeval))
-
-        # We are done
-        self.write_to_pipe(self.events)
-
-
-def quartz_mouse_process(pipe):
-    """Single subprocess for reading mouse events on Mac using newer Quartz."""
-    # Quartz only on the mac, so don't warn about Quartz
-    # pylint: disable=import-error
-    import Quartz
-
-    # pylint: disable=no-member
-
-    class QuartzMouseListener(QuartzMouseBaseListener):
-        """Loosely emulate Evdev mouse behaviour on the Macs.
-        Listen for key events then buffer them in a pipe.
-        """
-
-        def install_handle_input(self):
-            """Constants below listed at:
-            https://developer.apple.com/documentation/coregraphics/
-            cgeventtype?language=objc#topics
-            """
-            # Keep Mac Names to make it easy to find the documentation
-            # pylint: disable=invalid-name
-
-            NSMachPort = Quartz.CGEventTapCreate(
-                Quartz.kCGSessionEventTap,
-                Quartz.kCGHeadInsertEventTap,
-                Quartz.kCGEventTapOptionDefault,
-                Quartz.CGEventMaskBit(Quartz.kCGEventLeftMouseDown)
-                | Quartz.CGEventMaskBit(Quartz.kCGEventLeftMouseUp)
-                | Quartz.CGEventMaskBit(Quartz.kCGEventRightMouseDown)
-                | Quartz.CGEventMaskBit(Quartz.kCGEventRightMouseUp)
-                | Quartz.CGEventMaskBit(Quartz.kCGEventMouseMoved)
-                | Quartz.CGEventMaskBit(Quartz.kCGEventLeftMouseDragged)
-                | Quartz.CGEventMaskBit(Quartz.kCGEventRightMouseDragged)
-                | Quartz.CGEventMaskBit(Quartz.kCGEventScrollWheel)
-                | Quartz.CGEventMaskBit(Quartz.kCGEventTabletPointer)
-                | Quartz.CGEventMaskBit(Quartz.kCGEventTabletProximity)
-                | Quartz.CGEventMaskBit(Quartz.kCGEventOtherMouseDown)
-                | Quartz.CGEventMaskBit(Quartz.kCGEventOtherMouseUp)
-                | Quartz.CGEventMaskBit(Quartz.kCGEventOtherMouseDragged),
-                self.handle_input,
-                None,
-            )
-
-            CFRunLoopSourceRef = Quartz.CFMachPortCreateRunLoopSource(
-                None, NSMachPort, 0
-            )
-            CFRunLoopRef = Quartz.CFRunLoopGetCurrent()
-            Quartz.CFRunLoopAddSource(
-                CFRunLoopRef, CFRunLoopSourceRef, Quartz.kCFRunLoopDefaultMode
-            )
-            Quartz.CGEventTapEnable(NSMachPort, True)
-
-        def listen(self):
-            """Listen for quartz events."""
-            while self.active:
-                Quartz.CFRunLoopRunInMode(Quartz.kCFRunLoopDefaultMode, 5, False)
-
-        def uninstall_handle_input(self):
-            self.active = False
-
-        def _get_mouse_button_number(self, event):
-            """Get the mouse button number from an event."""
-            return Quartz.CGEventGetIntegerValueField(
-                event, Quartz.kCGMouseEventButtonNumber
-            )
-
-        def _get_click_state(self, event):
-            """The click state from an event."""
-            return Quartz.CGEventGetIntegerValueField(
-                event, Quartz.kCGMouseEventClickState
-            )
-
-        def _get_scroll(self, event):
-            """The scroll values from an event."""
-            scroll_y = Quartz.CGEventGetIntegerValueField(
-                event, Quartz.kCGScrollWheelEventDeltaAxis1
-            )
-            scroll_x = Quartz.CGEventGetIntegerValueField(
-                event, Quartz.kCGScrollWheelEventDeltaAxis2
-            )
-            return scroll_x, scroll_y
-
-        def _get_absolute(self, event):
-            """Get abolute cursor location."""
-            return Quartz.CGEventGetLocation(event)
-
-        def _get_relative(self, event):
-            """Get the relative mouse movement."""
-            delta_x = Quartz.CGEventGetIntegerValueField(
-                event, Quartz.kCGMouseEventDeltaX
-            )
-            delta_y = Quartz.CGEventGetIntegerValueField(
-                event, Quartz.kCGMouseEventDeltaY
-            )
-            return delta_x, delta_y
-
-    mouse = QuartzMouseListener(pipe)
-    mouse.listen()
-
-
-class AppKitMouseBaseListener(BaseListener):
-    """Emulate evdev behaviour on the the Mac."""
-
-    def __init__(self, pipe, events=None):
-        super(AppKitMouseBaseListener, self).__init__(
-            pipe, events, codes=dict(MAC_EVENT_CODES)
-        )
-
-    @staticmethod
-    def _get_mouse_button_number(event):
-        """Get the button number."""
-        return event.buttonNumber()
-
-    @staticmethod
-    def _get_absolute(event):
-        """Get the absolute (pixel) location of the mouse cursor."""
-        return event.locationInWindow()
-
-    @staticmethod
-    def _get_event_type(event):
-        """Get the appkit event type of the event."""
-        return event.type()
-
-    @staticmethod
-    def _get_deltas(event):
-        """Get the changes from the appkit event."""
-        delta_x = round(event.deltaX())
-        delta_y = round(event.deltaY())
-        delta_z = round(event.deltaZ())
-        return delta_x, delta_y, delta_z
-
-    def handle_button(self, event, event_type):
-        """Handle mouse click."""
-        mouse_button_number = self._get_mouse_button_number(event)
-        # Identify buttons 3,4,5
-        if event_type in (25, 26):
-            event_type = event_type + (mouse_button_number * 0.1)
-        # Add buttons to events
-        event_type_name, event_code, value, scan = self.codes[event_type]
-        if event_type_name == "Key":
-            scan_event, key_event = self.emulate_press(
-                event_code, scan, value, self.timeval
-            )
-            self.events.append(scan_event)
-            self.events.append(key_event)
-
-    def handle_absolute(self, event):
-        """Absolute mouse position on the screen."""
-        point = self._get_absolute(event)
-        x_pos = round(point.x)
-        y_pos = round(point.y)
-        x_event, y_event = self.emulate_abs(x_pos, y_pos, self.timeval)
-        self.events.append(x_event)
-        self.events.append(y_event)
-
-    def handle_scrollwheel(self, event):
-        """Make endev from appkit scroll wheel event."""
-        delta_x, delta_y, delta_z = self._get_deltas(event)
-        if delta_x:
-            self.events.append(self.emulate_wheel(delta_x, "x", self.timeval))
-        if delta_y:
-            self.events.append(self.emulate_wheel(delta_y, "y", self.timeval))
-        if delta_z:
-            self.events.append(self.emulate_wheel(delta_z, "z", self.timeval))
-
-    def handle_relative(self, event):
-        """Get the position of the mouse on the screen."""
-        delta_x, delta_y, delta_z = self._get_deltas(event)
-        if delta_x:
-            self.events.append(self.emulate_rel(0x00, delta_x, self.timeval))
-        if delta_y:
-            self.events.append(self.emulate_rel(0x01, delta_y, self.timeval))
-        if delta_z:
-            self.events.append(self.emulate_rel(0x02, delta_z, self.timeval))
-
-    def handle_input(self, event):
-        """Process the mouse event."""
-        self.update_timeval()
-        self.events = []
-        code = self._get_event_type(event)
-
-        # Deal with buttons
-        self.handle_button(event, code)
-
-        # Mouse wheel
-        if code == 22:
-            self.handle_scrollwheel(event)
-        # Other relative mouse movements
-        else:
-            self.handle_relative(event)
-
-        # Add in the absolute position of the mouse cursor
-        self.handle_absolute(event)
-
-        # End with a sync marker
-        self.events.append(self.sync_marker(self.timeval))
-
-        # We are done
-        self.write_to_pipe(self.events)
-
-
-def appkit_mouse_process(pipe):
-    """Single subprocess for reading mouse events on Mac using older AppKit."""
-    # pylint: disable=import-error,too-many-locals
-
-    # Note Objective C does not support a Unix style fork.
-    # So these imports have to be inside the child subprocess since
-    # otherwise the child process cannot use them.
-
-    # pylint: disable=no-member, no-name-in-module
-    from Foundation import NSObject
-    from AppKit import NSApplication, NSApp
-    from Cocoa import (
-        NSEvent,
-        NSLeftMouseDownMask,
-        NSLeftMouseUpMask,
-        NSRightMouseDownMask,
-        NSRightMouseUpMask,
-        NSMouseMovedMask,
-        NSLeftMouseDraggedMask,
-        NSRightMouseDraggedMask,
-        NSMouseEnteredMask,
-        NSMouseExitedMask,
-        NSScrollWheelMask,
-        NSOtherMouseDownMask,
-        NSOtherMouseUpMask,
-    )
-    from PyObjCTools import AppHelper
-    import objc
-
-    class MacMouseSetup(NSObject):
-        """Setup the handler."""
-
-        @objc.python_method
-        def init_with_handler(self, handler):
-            """
-            Init method that receives the write end of the pipe.
-            """
-            # ALWAYS call the super's designated initializer.
-            # Also, make sure to re-bind "self" just in case it
-            # returns something else!
-            # pylint: disable=self-cls-assignment
-            self = super(MacMouseSetup, self).init()
-            self.handler = handler
-            # Unlike Python's __init__, initializers MUST return self,
-            # because they are allowed to return any object!
-            return self
-
-        # pylint: disable=invalid-name, unused-argument
-        def applicationDidFinishLaunching_(self, notification):
-            """Bind the listen method as the handler for mouse events."""
-
-            mask = (
-                NSLeftMouseDownMask
-                | NSLeftMouseUpMask
-                | NSRightMouseDownMask
-                | NSRightMouseUpMask
-                | NSMouseMovedMask
-                | NSLeftMouseDraggedMask
-                | NSRightMouseDraggedMask
-                | NSScrollWheelMask
-                | NSMouseEnteredMask
-                | NSMouseExitedMask
-                | NSOtherMouseDownMask
-                | NSOtherMouseUpMask
-            )
-            NSEvent.addGlobalMonitorForEventsMatchingMask_handler_(mask, self.handler)
-
-    class MacMouseListener(AppKitMouseBaseListener):
-        """Loosely emulate Evdev mouse behaviour on the Macs.
-        Listen for key events then buffer them in a pipe.
-        """
-
-        def install_handle_input(self):
-            """Install the hook."""
-            self.app = NSApplication.sharedApplication()
-            # pylint: disable=no-member
-            delegate = MacMouseSetup.alloc().init_with_handler(self.handle_input)
-            NSApp().setDelegate_(delegate)
-            AppHelper.runEventLoop()
-
-        def __del__(self):
-            """Stop the listener on deletion."""
-            AppHelper.stopEventLoop()
-
-    # pylint: disable=unused-variable
-    mouse = MacMouseListener(pipe, events=[])
 
 
 class AppKitKeyboardListener(BaseListener):
@@ -899,53 +337,6 @@ class Keyboard(InputDevice):
             return super(Keyboard, self)._get_data(read_size)
         return self._pipe.recv_bytes()
 
-
-class Mouse(InputDevice):
-    """A mouse or other pointing-like device."""
-
-    def _set_device_path(self):
-        super(Mouse, self)._set_device_path()
-        if MAC:
-            self._device_path = APPKIT_MOUSE_PATH
-
-    def _set_name(self):
-        super(Mouse, self)._set_name()
-        if WIN:
-            self.name = "Microsoft Mouse"
-        elif MAC:
-            self.name = "AppKit Mouse"
-
-    @staticmethod
-    def _get_target_function():
-        """Get the correct target function."""
-        if WIN:
-            return mouse_process
-        if MAC:
-            return appkit_mouse_process
-        return None
-
-    def _get_data(self, read_size):
-        """Get data from the character device."""
-        if NIX:
-            return super(Mouse, self)._get_data(read_size)
-        return self._pipe.recv_bytes()
-
-
-class MightyMouse(Mouse):
-    """A mouse or other pointing device on the Mac."""
-
-    def _set_device_path(self):
-        super(MightyMouse, self)._set_device_path()
-        if MAC:
-            self._device_path = QUARTZ_MOUSE_PATH
-
-    def _set_name(self):
-        self.name = "Quartz Mouse"
-
-    @staticmethod
-    def _get_target_function():
-        """Get the correct target function."""
-        return quartz_mouse_process
